Java Programming
Create a class Student with the variables to represent the name of the student and marks of the five subjects store into an array, constructor to initialize the variables. Write method to compute total average marks, method to calculate letter grade, mutator and accessor methods, toString method to return object as a string. Create Driver class to demonstrate the functioning of the above.